readyup|refactor: Rename the project-discovery module for its two sweeps
`discoverKitProjects.ts` becomes `project-discovery.ts`, with its suite renamed alongside it. The module exports two production-consumed sweeps rather than one, so it takes a name describing its contents instead of one of the two exports.
