+++ eslint.config.ts
@@ -109,7 +109,7 @@ const config = defineConfig([
       'packages/readyup/src/list/__tests__/listCommand.recursive.unit.test.ts',
       'packages/readyup/src/list/__tests__/listCommand.recursivePackages.unit.test.ts',
       'packages/readyup/src/portable/__tests__/walkDirectories.unit.test.ts',
-      'packages/readyup/src/projects/__tests__/discoverKitProjects.unit.test.ts',
+      'packages/readyup/src/projects/__tests__/project-discovery.unit.test.ts',
       'packages/readyup/src/run/__tests__/resolveConfiguredPackages.unit.test.ts',
     ],
     rules: {
