+++ packages/readyup/src/projects/project-discovery.ts
